chore: moved lifi swapper into nested swapper directory
