+++ src/lib/swapper/swappers/LifiSwapper/utils/processGasCosts/processGasCosts.ts
@@ -2,7 +2,7 @@ import type { GasCost } from '@lifi/sdk'
 import type { Asset } from '@shapeshiftoss/asset-service'
 import type { BigNumber } from 'lib/bignumber/bignumber'
 import { bn, bnOrZero } from 'lib/bignumber/bignumber'
-import { getEvmAssetAddress } from 'lib/swapper/LifiSwapper/utils/getAssetAddress/getAssetAddress'
+import { getEvmAssetAddress } from 'lib/swapper/swappers/LifiSwapper/utils/getAssetAddress/getAssetAddress'
 
 const processNetworkFee = ({
   allRouteGasCosts,
